fix: cast index in ling int

+++ 0x1A-hash_tables/2-key_index.c
@@ -10,7 +10,7 @@
 
 unsigned long int key_index(const unsigned char *key, unsigned long int size)
 {
-	int index = hash_djb2(key);
+	unsigned long int index = hash_djb2(key);
 
 	return (index % size);
 }
